Replaces insertion sort and reverse with reverse order insertion sort
Whenever sortCellsByWealth() is called, reverse() is called immediately after. Sorting the cells in reverse order in the first place is more efficient. This commit removes the sortCellsByWealth() function, replaces it with sortCellsByWealthDescending() (which sorts in descending order), and replaces two instances of calls to sortCellsByWealth() and then reverse() with sortCellsByWealthDescending().