Fixes indentation and loop logic in findEthicalValueOfCell()
This change is purely based on the fact that `Egoist findEthicalValueOfCell()` is calling `neighbor` outside of the `for neighbor in self.neighborhood` loop, which doesn't seem right to me. I put the `canSee` sum in a separate loop because it seems like it's only supposed to be used in the main loop after it's fully calculated. `canSee` is used for calculating `extent` for each `neighbor`, but to be honest, I'm not sure what `extent` is doing inside of the loop. I thought it was indicated by the summing of each `neighborValueOfCell` into `cellValue`. `canSee` is one of the only differences between the `findEthicalValueOfCell()` in `Egoist` and the one in `Bentham`, which just uses `len(self.neighborhood)` to calculate `extent`.